requirements.txt
numpy
skyfield
plotly
flask
pyyaml
pytest
matplotlib 